Move core library to intgemm/ directory
Add proper LICENSE for catch.hpp
