+++ intgemm/multiply.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "intgemm_config.h"
+#include "intgemm/intgemm_config.h"
 #include "interleave.h"
 #include "intrinsics.h"
 #include "vec_traits.h"
