+++ intgemm/stats.inl
