+++ test/kernels/bitwise_not_test.cc
@@ -1,6 +1,6 @@
 #include "../test.h"
-#include "../../aligned.h"
-#include "../../kernels.h"
+#include "../../intgemm/aligned.h"
+#include "../../intgemm/kernels.h"
 
 #include <cstdlib>
 #include <numeric>
